backend/requirements.txt
fastapi
uvicorn
groq
httpx
python-dotenv
pandas